[GPU] Make Mali GPU driver compatible with our linux 3.5
This makes the possibility of OOM less likely with the binary hacked A800S GPU
driver that Henrick Grimler did back in 18.1, March 2021. The original mali
driver stopped working in 18.1.

However, the hacked A800S and r15p0 driver was prone to OOM with browsers and
Supercell games like Clash Royale would instanly crash.

Huge thanks to HazouPH for finding this commit that was in the original
17.1 r7p0 driver code, but never ported over. It took over 2 years
before HazouPH came up with this patch.

--- original commits by Henrick for A800S mali driver and r15p0 ---

https://github.com/exynos5420/android_vendor_samsung_universal5420-common/commit/e588b32699c59ce4bc661168f6e4ce7324c20d74

https://github.com/exynos5420/android_kernel_samsung_exynos5420/commit/ff87a1ef26d4d956e0d6554fb6ef8c4f1f3aa0fe